pyproject.toml
[tool.poetry]
name = "flac-to-vorbis"
version = "0.1.0"
description = ""
authors = ["Anthyme <90059514+derorsche@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^6.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
